GUI folder refactor
